accounts and transactions

+++ includes/delCart.php
@@ -5,19 +5,50 @@
 include("../includes/indexdb.php");
 include("../includes/session.php");
 
+function total(){
+  include("../includes/indexdb.php");
+include("../includes/session.php");
+
+  $totalPrice = 0;
+
+      $conn = new mysqli($servername, $username, $password, $dbname);
+
+      $sql = "SELECT sum(prodPrice) as total FROM tblcart where userId like '".$_SESSION['login_user']."'";
+      $result = $conn->query($sql);
+       
+     if ($result->num_rows > 0) {
+    // output data of each r1ow
+    
+    
+  
+        while($row = $result->fetch_assoc()) {
+              $totalPrice = $row["total"]; 
+        
+           
+    }
+
+    return $totalPrice;
+}
+}
         
     try {
        
                   
 
-                  $userid = $_SESSION['login_user'];
+                $userid = $_SESSION['login_user'];
+                $name = $_GET['name'];
+                $address = $_GET['address'];
+                $totalAmount = total();
               
 
 
                 $conn = new mysqli($servername, $username, $password, $dbname);
-         
-        
-                 $sql = "DELETE from tblcart where userId like '".$userid."'";
+
+                $sql = "INSERT INTO tbltransactions (userId, totalAmount, customerName, customerAddress) VALUES ('$userid', '$totalAmount' ,'$name','$address')";
+
+                  if($conn->query($sql) === TRUE) {
+
+                      $sql = "DELETE from tblcart where userId like '".$userid."'";
       
 
           
@@ -35,6 +66,14 @@ include("../includes/session.php");
                     echo "Order Failed";
                        
                       }
+
+                  }
+                  else{
+
+                  }
+         
+        
+
